+++ ckan/logic/action/get.py
@@ -27,6 +27,7 @@ import ckan.lib.search as search
 import ckan.lib.plugins as lib_plugins
 import ckan.lib.activity_streams as activity_streams
 import ckan.lib.datapreview as datapreview
+import ckan.lib.uploader as uploader
 import ckan.authz as authz
 
 from ckan.common import _
@@ -1168,6 +1169,44 @@ def resource_view_list(context, data_dict):
     ]
     return model_dictize.resource_view_list_dictize(resource_views, context)
 
+
+def resource_file_metadata_show(context, data_dict):
+    '''Get file metadata from a resource if it was uploaded.
+
+    :param id: the id of the resource
+    :type id: string
+
+    :returns: the meta data of resource file { 'content_type': content_type, 'size': length, 'hash': hash }
+    :rtype: dictionary
+    '''
+    model = context['model']
+    id = _get_or_bust(data_dict, 'id')
+
+    resource = model.Resource.get(id)
+    resource_context = dict(context, resource=resource)
+
+    if not resource:
+        raise NotFound
+
+    _check_access('resource_file_metadata_show', resource_context, data_dict)
+
+    pkg_dict = logic.get_action('package_show')(
+        dict(context),
+        {'id': resource.package.id,
+         'include_tracking': asbool(data_dict.get('include_tracking', False))})
+
+    for resource_dict in pkg_dict['resources']:
+        if resource_dict['id'] == id:
+            break
+    else:
+        log.error('Could not find resource %s after all', id)
+        raise NotFound(_('Resource was not found.'))
+
+    upload = uploader.get_resource_uploader(resource_dict)
+    try:
+        return upload.metadata(id)
+    except IOError:
+        raise NotFound
 
 @logic.auth_audit_exempt
 def revision_show(context, data_dict):