+++ ckan/tests/logic/action/test_get.py
@@ -1,11 +1,13 @@
 # encoding: utf-8
 
+import __builtin__ as builtins
 import datetime
 
 import nose.tools
 
 from six import text_type
 from six.moves import xrange
+import mock
 
 from ckan import __version__
 import ckan.logic as logic
@@ -14,11 +16,18 @@ import ckan.tests.helpers as helpers
 import ckan.tests.factories as factories
 import ckan.logic.schema as schema
 from ckan.lib.search.common import SearchError
+from pyfakefs import fake_filesystem
+
+from ckan.lib import uploader as ckan_uploader
 
 
 eq = nose.tools.eq_
 ok = nose.tools.ok_
 assert_raises = nose.tools.assert_raises
+
+fs = fake_filesystem.FakeFilesystem()
+fake_os = fake_filesystem.FakeOsModule(fs)
+fake_open = fake_filesystem.FakeFileOpen(fs)
 
 
 class TestPackageShow(helpers.FunctionalTestBase):
@@ -1822,6 +1831,62 @@ class TestShowResourceView(object):
         nose.tools.assert_raises(
             logic.NotFound,
             helpers.call_action, 'resource_view_show', id='does_not_exist')
+
+
+def mock_open_if_open_fails(*args, **kwargs):
+    try:
+        return real_open(*args, **kwargs)
+    except (OSError, IOError):
+        return fake_open(*args, **kwargs)
+
+
+class ShowResourceFileMetadata(object):
+
+    @helpers.change_config('ckan.storage_path', '/doesnt_exist')
+    @mock.patch.object(builtins, 'open', side_effect=mock_open_if_open_fails)
+    @mock.patch.object(ckan_uploader, 'os', fake_os)
+    @mock.patch.object(ckan_uploader, '_storage_path', new='/doesnt_exist')
+    def test_resource_file_metadata_show_meta_data_returned(self, _):
+        user = factories.User()
+        pkg = factories.Dataset(creator_user_id=user['id'])
+
+        url = url_for(
+            controller='api',
+            action='action',
+            logic_function='resource_create', ver='/3')
+        env = {'REMOTE_USER': user['name'].encode('ascii')}
+        postparams = {
+            'name': 'test-flask-upload',
+            'package_id': pkg['id']
+        }
+        upload_content = 'test-content,and2'
+        upload_info = ('upload', 'test-upload.csv', upload_content)
+        app = self._get_test_app()
+        resp = app.post(
+            url, params=postparams,
+            upload_files=[upload_info],
+            extra_environ=env
+        )
+        result = resp.json['result']
+        eq('upload', result['url_type'])
+        eq(len(upload_content), result['size'])
+
+        metaresult = helpers.call_action('resource_file_metadata_show', id=result['id'])
+        eq(len(upload_content), metaresult['size'])
+        eq('text/csv', metaresult['content_type'])
+        eq('', metaresult['hash'])
+
+    def test_resource_file_metadata_show_id_missing(self):
+
+        nose.tools.assert_raises(
+            logic.ValidationError,
+            helpers.call_action, 'resource_file_metadata_show')
+
+    def test_resource_file_metadata_show_id_not_found(self):
+
+        nose.tools.assert_raises(
+            logic.NotFound,
+            helpers.call_action, 'resource_file_metadata_show', id='does_not_exist')
 
 
 class TestGetHelpShow(object):